+++ Lib/fontbakery/profiles/fontbureau.py
@@ -4,6 +4,7 @@ PROFILE = {
     "pending_review": [
         "opentype/weight_class_fvar",
         "opentype/slant_direction",
+        "base_has_width",
         "cjk_not_enough_glyphs",
         "cmap/format_12",
         "color_cpal_brightness",